Add a utility package for decoding and encoding *certmanager.Certificate
